- Made DTOs and entities more nullable and consistent
- Added upload/update and filter/search DTOs
- Introduced search entities and new specifications
- Enhanced repository/service interfaces for search/filter
- Improved entity navigation and parent/child modeling
- Added AzureBlobService for file uploads
- Updated NuGet packages and reorganized namespaces
- Refactored Role entity and select list helpers
- Improved error handling and cleaned up unused properties

+++ Dtos/PlanModels/PlanmPartInfo.cs
@@ -24,14 +24,14 @@ namespace PMApplication.Dtos.PlanModels
         public int? Stock { get; set; }
         public int? Width { get; set; }
 
-        public int? PlanogramId { get; set; }
-        public int? ClusterId { get; set; }
-        public int? ScratchPadId { get; set; }
-        public int? PlanogramPartId { get; set; }
-        public int? ClusterPartId { get; set; }
-        public int? ParentPlanogramPartId { get; set; }
+        public long? PlanogramId { get; set; }
+        public long? ClusterId { get; set; }
+        public long? ScratchPadId { get; set; }
+        public long? PlanogramPartId { get; set; }
+        public long? ClusterPartId { get; set; }
+        public long? ParentPlanogramPartId { get; set; }
         public long? PlanogramShelfId { get; set; }
-        public int? ClusterShelfId { get; set; }
+        public long? ClusterShelfId { get; set; }
         public int? PlanogramColumnId { get; set; }
         //public string PlanxPartId { get; set; }
         //public string ParentPlanxPartId { get; set; }
